+++ Source/SwiftLintFrameworkTests/ConfigurationTests.swift
@@ -53,7 +53,7 @@ class ConfigurationTests: XCTestCase {
         XCTAssertEqual(config.included, [])
         XCTAssertEqual(config.excluded, [])
         XCTAssertEqual(config.reporter, "xcode")
-        XCTAssertEqual(config.reporterFromString.identifier, "xcode")
+        XCTAssertEqual(reporterFromString(config.reporter).identifier, "xcode")
     }
 
     func testWhitelistRules() {